behave differently in jenkins and locally when switching branches

+++ scripts/switchBranches.fish
@@ -1,15 +1,27 @@
 #!/usr/bin/env fish
-cd $INNERWORKDIR/ArangoDB
-and git clean -dfx
-and git checkout -- .
-and git fetch
-and git checkout $argv[1]
-and git reset --hard origin/$argv[1]
-and if test $ENTERPRISEEDITION = On
-  cd enterprise
-  and git clean -dfx
+
+if set -q JENKINS_HOME
+  cd $INNERWORKDIR/ArangoDB
   and git checkout -- .
   and git fetch
-  and git checkout $argv[2]
-  and git reset --hard origin/$argv[2]
+  and git checkout $argv[1]
+  and git reset --hard origin/$argv[1]
+  and git clean -fdx
+  and if test $ENTERPRISEEDITION = On
+    cd enterprise
+    and git checkout -- .
+    and git fetch
+    and git checkout $argv[2]
+    and git reset --hard origin/$argv[2]
+    and git clean -fdx
+  end
+else
+  cd $INNERWORKDIR/ArangoDB
+  and git checkout $argv[1]
+  and git pull
+  and if test $ENTERPRISEEDITION = On
+    cd enterprise
+    and git checkout $argv[2]
+    and git pull
+  end
 end
